refactor: consolidate all exports into a single file (index.js)
