+++ src/components/sponsors/sponsors.jsx
@@ -1,6 +1,6 @@
 import { Icon, SponsorsContainer, IconContainer, SponsorsSectionHeading } from './sponsors.styles'
 import Marquee from 'react-fast-marquee'
-import { SponsorsData } from '../../config/content/sponsorsData/sponsors'
+import { SponsorsData } from '../../config/index'
 
 export const Sponsors = () => {
   return (
