+++ src/config/content/leads/leads.js
@@ -1,34 +1,35 @@
-export const Lead = [
+export default [
   {
-    id: 1,
-    role: 'Convenor',
-    name: 'Rajdeep Garnayak',
+    role: 'Chair',
+    name: 'Aiswarya R. Behera',
     organization: 'India Student Regional conf. 2024',
     image:
-      'https://res.cloudinary.com/dhv234qct/image/upload/t_hc2/AICHE_SRC/team_members/Bookings.jpg'
+      'https://res.cloudinary.com/dhv234qct/image/upload/v1717692333/AICHE_SRC/team_members/SRC%20Chair.jpg'
   },
-  {
-    id: 2,
-    role: 'Co-Convenor',
-    name: 'Abhilash Sabat',
-    organization: 'India Student Regional conf. 2024',
-    image:
-      'https://res.cloudinary.com/dhv234qct/image/upload/v1717692332/AICHE_SRC/team_members/Event.jpg'
-  },
-  {
-    id: 3,
-    role: 'Content',
-    name: 'Umesh Chandra Sahoo',
-    organization: 'India Student Regional conf. 2024',
-    image:
-      'https://res.cloudinary.com/dhv234qct/image/upload/t_tchc1/AICHE_SRC/team_members/Accommodation.jpg'
-  }
+  [
+    {
+      id: 1,
+      role: 'Convenor',
+      name: 'Rajdeep Garnayak',
+      organization: 'India Student Regional conf. 2024',
+      image:
+        'https://res.cloudinary.com/dhv234qct/image/upload/t_hc2/AICHE_SRC/team_members/Bookings.jpg'
+    },
+    {
+      id: 2,
+      role: 'Convenor',
+      name: 'Abhilash Sabat',
+      organization: 'India Student Regional conf. 2024',
+      image:
+        'https://res.cloudinary.com/dhv234qct/image/upload/v1717692332/AICHE_SRC/team_members/Event.jpg'
+    },
+    {
+      id: 3,
+      role: 'Treasurer',
+      name: 'Umesh Chandra Sahoo',
+      organization: 'India Student Regional conf. 2024',
+      image:
+        'https://res.cloudinary.com/dhv234qct/image/upload/t_tchc1/AICHE_SRC/team_members/Accommodation.jpg'
+    }
+  ]
 ]
-
-export const SRCchair = {
-  role: 'Chair',
-  name: 'Aiswarya R. Behera',
-  organization: 'India Student Regional conf. 2024',
-  image:
-    'https://res.cloudinary.com/dhv234qct/image/upload/v1717692333/AICHE_SRC/team_members/SRC%20Chair.jpg'
-}
