+++ src/config/index.js
@@ -1,0 +1,21 @@
+// Desc: Content data for the website
+
+// Team members data
+export { default as TeamMembers } from './content/teamsData/team'
+
+// Sponsors data
+export { default as SponsorsData } from './content/sponsorsData/sponsors'
+
+// Registraion help data
+export { default as apostrophe } from './content/registrationHelp/registrationHelp'
+
+// Leads data
+export { default as LeadsData } from './content/leads/leads'
+
+// Nav data
+import { links, navLinks, DescriptionContent, heroContent } from './content/Nav_Hero'
+export { links, navLinks, DescriptionContent, heroContent }
+
+// Events data
+import { events, seeMoreIcon, XIcon, nextArrowIcon, prevArrowIcon } from './content/events/events'
+export { events, seeMoreIcon, XIcon, nextArrowIcon, prevArrowIcon }
